[arcane,core] Déplace la classe 'ItemEnumeratorBase' dans son propre fichier.
