+++ arcane/src/arcane/ItemEnumerator.h
@@ -18,6 +18,7 @@
 #include "arcane/Item.h"
 #include "arcane/EnumeratorTraceWrapper.h"
 #include "arcane/IItemEnumeratorTracer.h"
+#include "arcane/ItemEnumeratorBase.h"
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
@@ -40,148 +41,6 @@ namespace Arcane
 
 class ItemEnumeratorCS;
 class ItemGroupImpl;
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-/*!
- * \brief Classe de base des énumérateurs sur une liste d'entité.
- *
- * Les instances de cette classes sont créées soit via ItemEnumerator, soit
- * via ItemEnumeratorT.
- */
-template<typename ItemType>
-class ItemEnumeratorBaseT
-{
- private:
-
-  using ItemInternalPtr = ItemInternal*;
-  using LocalIdType = typename ItemType::LocalIdType;
-
- protected:
-
-  ItemEnumeratorBaseT()
-  : m_items(nullptr), m_local_ids(nullptr), m_index(0), m_count(0), m_group_impl(nullptr) { _init(); }
-  ItemEnumeratorBaseT(const ItemInternalPtr* items,const Int32* local_ids,Integer n, const ItemGroupImpl * agroup = nullptr)
-  : m_items(items), m_local_ids(local_ids), m_index(0), m_count(n), m_group_impl(agroup) { _init(); }
-  ItemEnumeratorBaseT(const ItemInternalArrayView& items,const Int32ConstArrayView& local_ids, const ItemGroupImpl * agroup = nullptr)
-  : m_items(items.data()), m_local_ids(local_ids.data()), m_index(0), m_count(local_ids.size()), m_group_impl(agroup) { _init(); }
-  ItemEnumeratorBaseT(const ItemInternalVectorView& view, const ItemGroupImpl * agroup = nullptr)
-  : m_items(view.items().data()), m_local_ids(view.localIds().data()),
-    m_index(0), m_count(view.size()), m_group_impl(agroup) { _init(); }
-  ItemEnumeratorBaseT(const ItemVectorView& rhs)
-  : ItemEnumeratorBaseT((const ItemInternalVectorView&)rhs) {}
-  ItemEnumeratorBaseT(const ItemVectorViewT<ItemType>& rhs)
-  : ItemEnumeratorBaseT((const ItemInternalVectorView&)rhs) {}
-
-  ItemEnumeratorBaseT(const ItemEnumerator& rhs);
-  ItemEnumeratorBaseT(const ItemInternalEnumerator& rhs);
-
- public:
-
-  //! Incrémente l'index de l'énumérateur
-  constexpr void operator++()
-  {
-    ++m_index;
-    m_is_not_end = (m_index<m_count);
-    if (m_is_not_end)
-      m_base.m_local_id = m_local_ids[m_index];
-  }
-  constexpr bool operator()() { return m_is_not_end; }
-
-  //! Vrai si on n'a pas atteint la fin de l'énumérateur (index()<count())
-  constexpr bool hasNext() { return m_is_not_end; }
-
-  //! Nombre d'éléments de l'énumérateur
-  constexpr Integer count() const { return m_count; }
-
-  //! Indice courant de l'énumérateur
-  constexpr Integer index() const { return m_index; }
-
-  //! localId() de l'entité courante.
-  constexpr Int32 itemLocalId() const { return m_base.m_local_id; }
-
-  //! localId() de l'entité courante.
-  constexpr Int32 localId() const { return m_base.m_local_id; }
-
-  /*!
-   * \internal
-   * \brief Indices locaux.
-   */
-  constexpr const Int32* unguardedLocalIds() const { return m_local_ids; }
-
-  /*!
-   * \internal
-   * \brief Liste des ItemInternal.
-   */
-  constexpr const ItemInternalPtr* unguardedItems() const { return m_items; }
-
-  /*!
-   * \internal
-   * \brief Partie interne (pour usage interne uniquement).
-   */
-  constexpr ItemInternal* internal() const { return m_items[m_base.m_local_id]; }
-
-  /*!
-   * \brief Groupe sous-jacent s'il existe (nullptr sinon)
-   *
-   * \brief Ceci vise à pouvoir tester que les accès par ce énumérateur sur un objet partiel sont licites.
-   */
-  constexpr const ItemGroupImpl* group() const { return m_group_impl; }
-
-  constexpr ItemType operator*() const { return m_base; }
-  constexpr const ItemType* operator->() const { return &m_base; }
-
-  constexpr LocalIdType asItemLocalId() const { return LocalIdType{m_base.m_local_id}; }
-
- public:
-
-  ItemEnumerator toItemEnumerator() const;
-
- protected:
-
-  // TODO Rendre privé
-  ItemType m_base;
-  const ItemInternalPtr* m_items;
-  const Int32* ARCANE_RESTRICT m_local_ids;
-  Int32 m_index;
-  Int32 m_count;
-  bool m_is_not_end;
-  const ItemGroupImpl* m_group_impl; // pourrait être retiré en mode release si nécessaire
-
- protected:
-
-  //! Constructeur seulement utilisé par fromItemEnumerator()
-  ItemEnumeratorBaseT(const ItemEnumerator& rhs,bool);
-
-  ItemEnumeratorBaseT(const ItemInternalPtr* items,const Int32* local_ids,Int32 index,Int32 n,
-                      const ItemGroupImpl * agroup,impl::ItemBase item_base)
-  : m_base(item_base), m_items(items), m_local_ids(local_ids), m_index(index), m_count(n), m_group_impl(agroup)
-  {
-    m_is_not_end = (m_index<m_count);
-  }
-
- private:
-
-  void _init()
-  {
-    m_is_not_end = (m_index<m_count);
-    if (m_is_not_end){
-      Int32 lid = m_local_ids[m_index];
-      // Vérifie qu'on n'indexe pas 'm_items' avec un localId() nul.
-      // Cela n'est pas possible avec les groupes d'entités mais ca l'est avec
-      // les ItemVector par exemple.
-      // De même, 'm_items' peut-être nul si tous les localId() sont nuls.
-      // Si c'est le cas, alors on prend comme ItemSharedInfo l'instance nulle.
-      // A terme le ItemSharedInfo sera renseigné directement dans le constructeur de
-      // cette classe.
-      Int32 idx = lid;
-      if (idx==NULL_ITEM_LOCAL_ID)
-        idx = 0;
-      ItemSharedInfo* isi = (m_items) ? m_items[idx]->sharedInfo() : ItemSharedInfo::nullInstance();
-      m_base = ItemType(ItemBaseBuildInfo(lid,isi));
-    }
-  }
-};
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
